Identity: Extract enum for Epoxy Items
